requirements.txt
openai>=1.0.0
pyyaml>=6.0
jupyter>=1.0.0
ipywidgets>=8.0.0
tabulate>=0.9.0
python-dotenv>=1.0.0
